requirements.txt
Django==1.5.1
psycopg2==2.5
bootstrap-admin==0.2.1
django-ajax-selects==1.2.5
